requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.11.0
asttokens==3.0.0
async-lru==2.0.5
asyncer==0.0.11
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
certifi==2025.7.14
cffi==2.0.0
chainlit==2.9.3
charset-normalizer==3.4.2
chevron==0.14.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
cuid==0.4
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.0
fastapi==0.124.0
fastjsonschema==2.21.1
filelock==3.20.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==0.37.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-groq==1.1.0
langchain-google-genai==4.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.15
langsmith==0.4.58
lark==1.2.2
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.23.3
mistune==3.1.3
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-agno==0.49.7
opentelemetry-instrumentation-alephalpha==0.49.7
opentelemetry-instrumentation-anthropic==0.49.7
opentelemetry-instrumentation-bedrock==0.49.7
opentelemetry-instrumentation-chromadb==0.49.7
opentelemetry-instrumentation-cohere==0.49.7
opentelemetry-instrumentation-crewai==0.49.7
opentelemetry-instrumentation-google-generativeai==0.49.7
opentelemetry-instrumentation-groq==0.49.7
opentelemetry-instrumentation-haystack==0.49.7
opentelemetry-instrumentation-lancedb==0.49.7
opentelemetry-instrumentation-langchain==0.49.7
opentelemetry-instrumentation-llamaindex==0.49.7
opentelemetry-instrumentation-logging==0.60b0
opentelemetry-instrumentation-marqo==0.49.7
opentelemetry-instrumentation-mcp==0.49.7
opentelemetry-instrumentation-milvus==0.49.7
opentelemetry-instrumentation-mistralai==0.49.7
opentelemetry-instrumentation-ollama==0.49.7
opentelemetry-instrumentation-openai==0.49.7
opentelemetry-instrumentation-openai-agents==0.49.7
opentelemetry-instrumentation-pinecone==0.49.7
opentelemetry-instrumentation-qdrant==0.49.7
opentelemetry-instrumentation-redis==0.60b0
opentelemetry-instrumentation-replicate==0.49.7
opentelemetry-instrumentation-requests==0.60b0
opentelemetry-instrumentation-sagemaker==0.49.7
opentelemetry-instrumentation-sqlalchemy==0.60b0
opentelemetry-instrumentation-threading==0.60b0
opentelemetry-instrumentation-together==0.49.7
opentelemetry-instrumentation-transformers==0.49.7
opentelemetry-instrumentation-urllib3==0.60b0
opentelemetry-instrumentation-vertexai==0.49.7
opentelemetry-instrumentation-watsonx==0.49.7
opentelemetry-instrumentation-weaviate==0.49.7
opentelemetry-instrumentation-writer==0.49.7
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b0
orjson==3.11.5
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
platformdirs==4.3.8
portalocker==3.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.33.2
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.15.0

PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.16.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
Send2Trash==1.8.3
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.45
sqlite-vec==0.1.6
sqlparse==0.5.4
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
torch
tornado==6.5.1
tqdm==4.67.1
traceloop-sdk==0.49.7
traitlets==5.14.3
transformers==4.57.3
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.38.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
elevenlabs==2.27.0


Pillow
huggingface_hub
